+++ tests/test_column_loading.py
@@ -22,3 +22,15 @@ class TestColumnReflection(TestCase):
             comment='test column'
         )
         assert isinstance(column_info_1['type'], VARCHAR)
+
+    def test_encode_in_dialect_options(self):
+        """
+        Column encoding should be a dialect_kwarg under dialect_options
+        """
+        dialect = RedshiftDialect()
+        column_info = dialect._get_column_info(
+            'Column with encoding',
+            'integer', None, False, {}, {}, 'default', encode='LZO',
+            comment='test encoding column'
+        )
+        assert column_info['dialect_options']['redshift_encode'] == 'LZO'